% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACOSOG.genotype.list}
\alias{ACOSOG.genotype.list}
\title{ACOSOG.genotype.list-compiled TCGA BRCA CPM log2 matrix}
\format{
A list
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
ACOSOG.genotype.list
}
\description{
A data set
}
\keyword{datasets}
