% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACOSOG.phenoData}
\alias{ACOSOG.phenoData}
\title{ACOSOG.phenoData}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
ACOSOG.phenoData
}
\description{
A data set
}
\keyword{datasets}
