% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACOSOG.preCalmatrix}
\alias{ACOSOG.preCalmatrix}
\title{ACOSOG.preCalmatrix}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
ACOSOG.preCalmatrix
}
\description{
A data set
}
\keyword{datasets}
