% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{batch_CalWeight_Similarity}
\alias{batch_CalWeight_Similarity}
\title{Title batch calculate weight for a target list and a compendia of comparing lists}
\usage{
batch_CalWeight_Similarity(
  target.list,
  compare.list,
  trainset = NULL,
  method = c("logistic.regression", "pearson", "phi", "ochiai", "jaccard"),
  minsize = 10,
  confound.factor = NULL
)
}
\arguments{
\item{target.list}{target samples to calculate iGenSig score}

\item{compare.list}{compare genotype list}

\item{trainset}{trainset data}

\item{method}{method for similarity calculation}

\item{minsize}{minimum size}

\item{confound.factor}{confounding factors}
}
\value{
target result
}
\description{
Title batch calculate weight for a target list and a compendia of comparing lists
}
\examples{
batch_CalWeight_Similarity(target.list,compare.list,trainset=NULL,method=c("logistic.regression","pearson","phi","ochiai","jaccard"),minsize=10,confound.factor=NULL)
}
