% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{batchbenchmark.dGenSig.trial}
\alias{batchbenchmark.dGenSig.trial}
\title{Title benchmark dGenSig scores based on binary response data and/or survival data}
\usage{
batchbenchmark.dGenSig.trial(
  gensig.dir,
  phenoData,
  event.col = NULL,
  time.col = NULL,
  by.cluster,
  validationset = FALSE
)
}
\arguments{
\item{gensig.dir}{CALGB.gensigdir}

\item{phenoData}{CALGB phenodata}

\item{event.col}{survival event data}

\item{time.col}{survival time data}

\item{by.cluster}{clustering method}

\item{validationset}{validation dataset name}
}
\value{
phenoData.xls
}
\description{
Title benchmark dGenSig scores based on binary response data and/or survival data
}
\examples{
batchbenchmark.dGenSig.trial(gensig.dir, phenoData,event.col=NULL,time.col=NULL,by.cluster,validationset=FALSE)
}
