% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{batchCal.dGenSig.trial2trial.train}
\alias{batchCal.dGenSig.trial2trial.train}
\title{Title Batch calculate dGenSig score based on training clinical trial dataset}
\usage{
batchCal.dGenSig.trial2trial.train(
  gensigdir,
  test.matrix = NULL,
  phenoData,
  method = c("math", "ridge")
)
}
\arguments{
\item{gensigdir}{iGenSig-Rx output directory}

\item{test.matrix}{testset matrix}

\item{phenoData}{testset phenotype data}

\item{method}{dGenSig calculation method}
}
\value{
return value is at cal.dGenSig.trial2trial.math
}
\description{
Title Batch calculate dGenSig score based on training clinical trial dataset
}
\examples{
batchCal.dGenSig.trial2trial.train(gensigdir=CALGB.gensigdir,test.matrix=NULL,phenoData,method=c("math","ridge"))
}
