% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{batchCal.dGenSig.trial2trial.validation}
\alias{batchCal.dGenSig.trial2trial.validation}
\title{Title batch calculation of dGenSig in validation data}
\usage{
batchCal.dGenSig.trial2trial.validation(
  train.gensigdir,
  validation.gensigdir,
  method = c("math", "ridge")
)
}
\arguments{
\item{train.gensigdir}{iGenSig-Rx directory that has train data}

\item{validation.gensigdir}{validation data gensig directory}

\item{method}{method to calculate dGenSig score}
}
\value{
it will write result.plot in cal.dGenSig.trial2trial.math.validation(0)
}
\description{
Title batch calculation of dGenSig in validation data
}
\examples{
batchCal.dGenSig.trial2trial.validation(train.gensigdir,validation.gensigdir,method=c("math","ridge"))
}
