% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{batchCal.GenSig.similarity}
\alias{batchCal.GenSig.similarity}
\title{Title batch calculate uniGenSig scores based on preCalculated weights. subject.sensitive must only contain subjects in the training set}
\usage{
batchCal.GenSig.similarity(
  subject.sensitive = NULL,
  feature.select = NULL,
  weightfile = NULL,
  trainset = NULL,
  outprefix,
  subject.genotype.list,
  preCalmatrix = NULL,
  tcga.genotype.list = NULL,
  subject.gensigdir,
  method = c("logistic.regression", "pearson", "phi", "ochiai", "jaccard"),
  redun.method,
  redun.cut,
  minsize = 10,
  sen.weightcut,
  res.weightcut,
  sen.pcut = NULL,
  res.pcut = NULL,
  power = 2,
  root = 0.5,
  ECNpenalty = 1,
  rm.equivocal,
  highlevel.minsize = 10,
  confound.factor = NULL,
  by.cluster = FALSE,
  validationset = FALSE
)
}
\arguments{
\item{subject.sensitive}{subject IDs of drug sensitive}

\item{feature.select}{selected features}

\item{weightfile}{weight file name}

\item{trainset}{trainset name}

\item{outprefix}{outfile prefix}

\item{subject.genotype.list}{CALGB genotype list}

\item{preCalmatrix}{TCGA preCalmatrix}

\item{tcga.genotype.list}{TCGA genotype list}

\item{subject.gensigdir}{CALGB iGenSig output directory}

\item{method}{weight calculation method}

\item{redun.method}{redundancy method}

\item{redun.cut}{redundancy cutoff}

\item{minsize}{minimum size of features}

\item{sen.weightcut}{sensitive score weight cutoff}

\item{res.weightcut}{resistant score weight cutoff}

\item{sen.pcut}{sensitive score p cutoff}

\item{res.pcut}{resistant score p cutoff}

\item{power}{power in iGenSig-Rx formula}

\item{root}{root in iGenSig-Rx formula}

\item{ECNpenalty}{ECNpenalty in iGenSig-Rx formula}

\item{rm.equivocal}{removing equivocal features}

\item{highlevel.minsize}{minimum size of high level features}

\item{confound.factor}{confounding factors}

\item{by.cluster}{cluster methods to remove redundant features}

\item{validationset}{validation data}
}
\value{
iGenSig.xls file.
}
\description{
Title batch calculate uniGenSig scores based on preCalculated weights. subject.sensitive must only contain subjects in the training set
}
\examples{
batchCal.GenSig.similarity(subject.sensitive=NULL,feature.select=NULL,weightfile=NULL,trainset=NULL,outprefix,subject.genotype.list,preCalmatrix=NULL,tcga.genotype.list=NULL,subject.gensigdir,method=c("logistic.regression","pearson","phi","ochiai","jaccard"),redun.method,redun.cut,minsize=10,sen.weightcut,res.weightcut,sen.pcut=NULL,res.pcut=NULL,power=2,root=0.5,ECNpenalty=1,rm.equivocal,highlevel.minsize=10,confound.factor=NULL,by.cluster=FALSE,validationset=FALSE)
}
