% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{by.cluster}
\alias{by.cluster}
\title{by.cluster}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
by.cluster
}
\description{
A data set
}
\keyword{datasets}
