% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{cal.dGenSig.trial2trial.math}
\alias{cal.dGenSig.trial2trial.math}
\title{Title calculate dGenSig score based on linear method}
\usage{
cal.dGenSig.trial2trial.math(
  gensig.file,
  dgensig.file,
  subject.train = NULL,
  phenoData
)
}
\arguments{
\item{gensig.file}{iGenSig-Rx file}

\item{dgensig.file}{dGenSig file}

\item{subject.train}{trainset subject}

\item{phenoData}{phenotyp data}
}
\value{
write txt with result plot
}
\description{
Title calculate dGenSig score based on linear method
}
\examples{
cal.dGenSig.trial2trial.math(gensig.file, dgensig.file, subject.train=NULL, phenoData)
}
