% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{cal.dGenSig.trial2trial.math.validation}
\alias{cal.dGenSig.trial2trial.math.validation}
\title{Title calculate dGenSig in validation set}
\usage{
cal.dGenSig.trial2trial.math.validation(gensig.file, dgensig.file, mySlope)
}
\arguments{
\item{gensig.file}{input iGenSig-Rx file.}

\item{dgensig.file}{input dGenSig file}

\item{mySlope}{dline slope}
}
\value{
return value is from cal.dGenSig.trial2trial.ml.validation
}
\description{
Title calculate dGenSig in validation set
}
\examples{
cal.dGenSig.trial2trial.math.validation(gensig.file, dgensig.file, mySlope)
}
