% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{calCorrelation}
\alias{calCorrelation}
\title{Title calculate weight based on similarity or correlation statistics}
\usage{
calCorrelation(
  target.list,
  compare.list,
  list.all,
  confound.factor = NULL,
  confound.cut = 0.4
)
}
\arguments{
\item{target.list}{target samples in list}

\item{compare.list}{compare genotype list}

\item{list.all}{all list}

\item{confound.factor}{confounding factor}

\item{confound.cut}{confounding cutoff}
}
\value{
Weight based on similarity or correlation statistics.
}
\description{
Title calculate weight based on similarity or correlation statistics
}
\examples{
calCorrelation(target.list,compare.list,list.all,confound.factor=NULL,confound.cut=0.4)
}
