% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CALGB.genotype.matrix_Example}
\alias{CALGB.genotype.matrix_Example}
\title{CALGB.genotype.matrix_Example}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
CALGB.genotype.matrix_Example
}
\description{
A data set
}
\keyword{datasets}
