% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{calLogReg}
\alias{calLogReg}
\title{Title calculate weight based on logistic regression statistics with consideration of confound factor}
\usage{
calLogReg(
  target.list,
  compare.list,
  list.all,
  confound.factor = NULL,
  confound.pcut = 0.01
)
}
\arguments{
\item{target.list}{target samples to calculate iGenSig score}

\item{compare.list}{compare genotype list}

\item{list.all}{all list}

\item{confound.factor}{confounding factors}

\item{confound.pcut}{confounding p cutoff}
}
\value{
weight calculated based on logistic regression
}
\description{
Title calculate weight based on logistic regression statistics with consideration of confound factor
}
\examples{
calLogReg(target.list,compare.list,list.all,confound.factor=NULL,confound.pcut=0.01)
}
