% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{CalWeight.Similarity}
\alias{CalWeight.Similarity}
\title{Title Calculate weight based on ochiai Index}
\usage{
CalWeight.Similarity(list1, list2, list.all, method)
}
\arguments{
\item{list1}{1st list}

\item{list2}{2nd list}

\item{list.all}{all list}

\item{method}{similarity calculation method}
}
\value{
temporary weight
}
\description{
Title Calculate weight based on ochiai Index
}
\examples{
CalWeight.Similarity(list1,list2,list.all,method)
}
