% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compare.list}
\alias{compare.list}
\title{compare.list}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
compare.list
}
\description{
A data set
}
\keyword{datasets}
