% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{confound}
\alias{confound}
\title{confound}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
confound
}
\description{
A data set
}
\keyword{datasets}
