% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dgensig.method}
\alias{dgensig.method}
\title{dgensig.method}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
dgensig.method
}
\description{
A data set
}
\keyword{datasets}
