% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{FilterFeature}
\alias{FilterFeature}
\title{Title remove equivocal features in the filtered feature weight table. sig.feature.weight must contain Feature.List column. feature.list.train must only contain the subject used in training sets}
\usage{
FilterFeature(
  feature.weight,
  feature.list.train,
  weight.cut,
  low.weight.cut,
  p.cut = NULL,
  low.p.cut,
  feature.col = 1,
  highlevel.minsize = 10,
  filter.criteria = c(1, 2)
)
}
\arguments{
\item{feature.weight}{feature weight}

\item{feature.list.train}{feature list to train}

\item{weight.cut}{weight cutoff}

\item{low.weight.cut}{low weight cutoff}

\item{p.cut}{p cutoff}

\item{low.p.cut}{low p cutoff}

\item{feature.col}{the column number of features}

\item{highlevel.minsize}{the minimum size of features}

\item{filter.criteria}{filter criteria}
}
\value{
feature weight
}
\description{
Title remove equivocal features in the filtered feature weight table. sig.feature.weight must contain Feature.List column. feature.list.train must only contain the subject used in training sets
}
\examples{
FilterFeature(feature.weight,feature.list.train,weight.cut,low.weight.cut,p.cut=NULL,low.p.cut,feature.col=1,highlevel.minsize=10,filter.criteria=c(1,2))
}
