% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{FilterFunction}
\alias{FilterFunction}
\title{Title FilterFunction by one list}
\usage{
FilterFunction(compare.list, DataToFilter, inlist = TRUE)
}
\arguments{
\item{compare.list}{compare.list}

\item{DataToFilter}{trainset subjects}

\item{inlist}{TRUE means filtering in.}
}
\value{
filtered target subjects
}
\description{
Title FilterFunction by one list
}
\examples{
FilterFunction(compare.list, DataToFilter, inlist=TRUE)
}
