% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{find.dline}
\alias{find.dline}
\title{Title find division line based on training set}
\usage{
find.dline(gensig.result)
}
\arguments{
\item{gensig.result}{iGenSig-Rx result data.}
}
\value{
slope value
}
\description{
Title find division line based on training set
}
\examples{
find.dline(gensig.result)
}
