% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fold.assign}
\alias{fold.assign}
\title{fold.assign}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
fold.assign
}
\description{
A data set
}
\keyword{datasets}
