% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{folder.prefix}
\alias{folder.prefix}
\title{folder.prefix}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
folder.prefix
}
\description{
A data set
}
\keyword{datasets}
