% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{genolist2matrix}
\alias{genolist2matrix}
\title{Title turn genotype list into a matrix fast}
\usage{
genolist2matrix(genolist)
}
\arguments{
\item{genolist}{}
}
\value{
genotype matrix
}
\description{
Title turn genotype list into a matrix fast
}
\examples{
genolist2matrix(CALGB.genotype.list_Example)
}
