% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genSig.outprefix}
\alias{genSig.outprefix}
\title{genSig.outprefix}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
genSig.outprefix
}
\description{
A data set
}
\keyword{datasets}
