% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list2}
\alias{list2}
\title{list2}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
list2
}
\description{
A data set
}
\keyword{datasets}
