% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{matrix2genofile}
\alias{matrix2genofile}
\title{Title turn GenotypeMatrix into genolist file}
\usage{
matrix2genofile(GenotypeMatrix, outfile)
}
\arguments{
\item{GenotypeMatrix}{genotype data in matrix class}

\item{outfile}{outfile name}
}
\value{
genotype file in gmt format
}
\description{
Title turn GenotypeMatrix into genolist file
}
\examples{
matrix2genofile(GenotypeMatrix=CALGB.genotype.matrix_Example, outfile="CALGB.genotype.gmt")
}
