% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minsize}
\alias{minsize}
\title{minsize}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
minsize
}
\description{
A data set
}
\keyword{datasets}
