% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{normalize}
\alias{normalize}
\title{Title scale the max of a vector to 1 and min to 0}
\usage{
normalize(NumbVector = NumbVector)
}
\arguments{
\item{NumbVector}{}
}
\value{
Normalized vector
}
\description{
Title scale the max of a vector to 1 and min to 0
}
\examples{
normalize(NumbVector=NumbVector)
}
