% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NumbVector}
\alias{NumbVector}
\title{NumbVector}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
NumbVector
}
\description{
A data set
}
\keyword{datasets}
