% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{read_gmt}
\alias{read_gmt}
\title{Title}
\usage{
read_gmt(fileName, min = 5)
}
\arguments{
\item{fileName}{CALGB gmt file.}

\item{min}{minimun number of samples per genotype}
}
\value{
genotype.list
}
\description{
Title
}
\examples{
read_gmt(CALGB.genotype.file, min=5)
}
