% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redun.method}
\alias{redun.method}
\title{redun.method}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
redun.method
}
\description{
A data set
}
\keyword{datasets}
