% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res.pcut}
\alias{res.pcut}
\title{res.pcut}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
res.pcut
}
\description{
A data set
}
\keyword{datasets}
