% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res.weightcut}
\alias{res.weightcut}
\title{res.weightcut}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
res.weightcut
}
\description{
A data set
}
\keyword{datasets}
