% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rm.equivocal}
\alias{rm.equivocal}
\title{rm.equivocal}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
rm.equivocal
}
\description{
A data set
}
\keyword{datasets}
