% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{simulate.seqError}
\alias{simulate.seqError}
\title{Title simulate sequencing errors in gmt files}
\usage{
\method{simulate}{seqError}(genotypefile, outpath, errPerc = 0.05, permuN = 5)
}
\arguments{
\item{genotypefile}{file name of gneotype list}

\item{outpath}{to get output file}

\item{errPerc}{percentage of error}

\item{permuN}{permutation number}
}
\value{
Error simulated genotype gmt file.
}
\description{
Title simulate sequencing errors in gmt files
}
\examples{
simulate.seqError(genotypefile=CALGB.genotype.file, outpath=".", errPerc=0.05, permuN=5)
}
