% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGenSigRx_Modules.R
\name{survival.analysis.trial}
\alias{survival.analysis.trial}
\title{Title Survival analysis module for dGenSig}
\usage{
survival.analysis.trial(
  dGenSig,
  dGenSig.file,
  event.col = 8,
  time.col = 9,
  rho = 0
)
}
\arguments{
\item{dGenSig}{result.plot}

\item{dGenSig.file}{output file}

\item{event.col}{survival event data}

\item{time.col}{survival time data}

\item{rho}{0}
}
\value{
survival analysis result
}
\description{
Title Survival analysis module for dGenSig
}
\examples{
survival.analysis.trial(dGenSig=result.plot[result.plot$TEST_SET==1,],dGenSig.file=dGenSig.file,event.col=event.col+basecol-1,time.col=time.col+basecol-1, rho=0)
}
