% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmp.weight}
\alias{tmp.weight}
\title{tmp.weight}
\format{
A matrix format.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/iGenSig-Rx}
}
\usage{
tmp.weight
}
\description{
A data set
}
\keyword{datasets}
